ESP32 CODE for Child Rescue from Open Borewell System using Robot
 The Child Rescue System uses a robotic arm with ultrasonic sensor and camera controlled by 
a microcontroller ESP32 and ESP32 cam module to safely extract children from open 
borewells. The remote control capability which is controlled by the blynk software and real
time video monitoring enhance rescue efficiency and safety, eliminating the need for manual 
intervention in risky environments and reducing human error.Also the readings of the 
ultrasonic sensor is shown in the blynk software as well. This innovative technology is a 
crucial step toward safeguarding children's lives and preventing fatalities in such incidents